packages/bot-event-dispatcher/requirements.txt
avro==1.11.1 ; python_version >= "3.10" and python_version < "4.0"
black==22.8.0 ; python_version >= "3.10" and python_version < "4.0"
cachetools==5.2.0 ; python_version >= "3.10" and python_version < "4.0"
certifi==2022.9.24 ; python_version >= "3.10" and python_version < "4"
charset-normalizer==2.1.1 ; python_version >= "3.10" and python_version < "4"
click==8.1.3 ; python_version >= "3.10" and python_version < "4.0"
cloudevents==1.6.1 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.5 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
deprecation==2.1.0 ; python_version >= "3.10" and python_version < "4.0"
flake8==5.0.4 ; python_version >= "3.10" and python_version < "4.0"
flask==2.2.2 ; python_version >= "3.10" and python_version < "4"
functions-framework==3.2.0 ; python_version >= "3.10" and python_version < "4"
google-api-core[grpc]==2.10.1 ; python_version >= "3.10" and python_version < "4.0"
google-auth==2.12.0 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-pubsub==2.13.7 ; python_version >= "3.10" and python_version < "4.0"
googleapis-common-protos==1.56.4 ; python_version >= "3.10" and python_version < "4.0"
googleapis-common-protos[grpc]==1.56.4 ; python_version >= "3.10" and python_version < "4.0"
grpc-google-iam-v1==0.12.4 ; python_version >= "3.10" and python_version < "4.0"
grpcio-status==1.49.1 ; python_version >= "3.10" and python_version < "4.0"
grpcio==1.49.1 ; python_version >= "3.10" and python_version < "4.0"
gunicorn==20.1.0 ; python_version >= "3.10" and python_version < "4" and platform_system != "Windows"
idna==3.4 ; python_version >= "3.10" and python_version < "4"
itsdangerous==2.1.2 ; python_version >= "3.10" and python_version < "4"
jinja2==3.1.2 ; python_version >= "3.10" and python_version < "4"
markupsafe==2.1.1 ; python_version >= "3.10" and python_version < "4"
mccabe==0.7.0 ; python_version >= "3.10" and python_version < "4.0"
mypy-extensions==0.4.3 ; python_version >= "3.10" and python_version < "4.0"
packaging==21.3 ; python_version >= "3.10" and python_version < "4.0"
pathspec==0.10.1 ; python_version >= "3.10" and python_version < "4.0"
platformdirs==2.5.2 ; python_version >= "3.10" and python_version < "4.0"
proto-plus==1.22.1 ; python_version >= "3.10" and python_version < "4.0"
protobuf==4.21.7 ; python_version >= "3.10" and python_version < "4.0"
pyasn1-modules==0.2.8 ; python_version >= "3.10" and python_version < "4.0"
pyasn1==0.4.8 ; python_version >= "3.10" and python_version < "4.0"
pycodestyle==2.9.1 ; python_version >= "3.10" and python_version < "4.0"
pyflakes==2.5.0 ; python_version >= "3.10" and python_version < "4.0"
pyparsing==3.0.9 ; python_version >= "3.10" and python_version < "4.0"
pytelegrambotapi==4.7.0 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==0.21.0 ; python_version >= "3.10" and python_version < "4.0"
requests==2.28.1 ; python_version >= "3.10" and python_version < "4"
rsa==4.9 ; python_version >= "3.10" and python_version < "4"
setuptools==65.4.1 ; python_version >= "3.10" and python_version < "4" and platform_system != "Windows"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
tomli==2.0.1 ; python_version >= "3.10" and python_full_version < "3.11.0a7"
ujson==5.5.0 ; python_version >= "3.10" and python_version < "4.0"
urllib3==1.26.12 ; python_version >= "3.10" and python_version < "4"
watchdog==2.1.9 ; python_version >= "3.10" and python_version < "4"
werkzeug==2.2.2 ; python_version >= "3.10" and python_version < "4"
